+++ app/components/vm-component.js
@@ -2,9 +2,9 @@
 import { useEffect } from 'react';
 import { useInitNear, Widget, EthersProviderContext } from 'near-social-vm';
 
-import { useWallet } from '@/wallets/wallet-selector';
-import { useEthersProviderContext } from '@/wallets/web3-wallet';
-import { NetworkId } from '@/config';
+import { useWallet } from '@/app/wallets/wallet-selector';
+import { useEthersProviderContext } from '@/app/wallets/web3-wallet';
+import { NetworkId } from '@/app/config';
 
 export default function Component({ src }) {
   const ethersContext = useEthersProviderContext();
