+++ app/page.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function Home() {
+  return (
+    <main className='h-screen bg-white'>
+      
+    </main>
+  );
+}
