--- src/app/pages/home.jsx
@@ -1,6 +1,0 @@
- import React from "react";
- import ResponsiveAppBar from "../components/homeNavbar";
-
-
- <ResponsiveAppBar/>
-
